+++ taskflowbackend/tasks/irods_tasks.py
@@ -21,6 +21,17 @@ from taskflowbackend.tasks.base_task import BaseTask
 logger = logging.getLogger('__name__')
 
 
+# Local constants
+# NOTE: This is only compabitle with iRODS 4.3.
+# Backwards compatibility with 4.2 has been removed in SODAR v1.0.
+ACCESS_LOOKUP = {
+    'read': 'read_object',
+    'read_object': 'read',
+    'write': 'modify_object',
+    'modify_object': 'write',
+    'null': 'null',
+    'own': 'own',
+}
 INHERIT_STRINGS = {True: 'inherit', False: 'noinherit'}
 META_EMPTY_VALUE = 'N/A'
 MD5_RE = re.compile(r'([^\w.])')
@@ -38,7 +49,6 @@ class IrodsAccessMixin:
         access_name,
         path,
         user_name,
-        access_lookup,
         obj_target,
         recursive,
     ):
@@ -48,7 +58,6 @@ class IrodsAccessMixin:
         :param access_name: Access level to set (string)
         :param path: Full iRODS path to collection or data object (string)
         :param user_name: Name of user or group (string)
-        :param access_lookup: Access level lookup for iRODS compatibility (dict)
         :param obj_target: Whether target is a data object (boolean)
         :param recursive: Set collection access recursively if True (boolean)
         """
@@ -68,9 +77,9 @@ class IrodsAccessMixin:
         modifying_data = False
         if (
             user_access
-            and user_access.access_name != access_lookup[access_name]
+            and user_access.access_name != ACCESS_LOOKUP[access_name]
         ):
-            self.execute_data['access_names'][path] = access_lookup[
+            self.execute_data['access_names'][path] = ACCESS_LOOKUP[
                 user_access.access_name
             ]
             modifying_data = True
@@ -379,7 +388,6 @@ class SetAccessTask(IrodsAccessMixin, IrodsBaseTask):
         access_name,
         path,
         user_name,
-        access_lookup,
         irods_backend,
         obj_target=False,
         recursive=True,
@@ -391,7 +399,6 @@ class SetAccessTask(IrodsAccessMixin, IrodsBaseTask):
                 access_name,
                 path,
                 user_name,
-                access_lookup,
                 obj_target,
                 recursive,
             )
@@ -404,7 +411,6 @@ class SetAccessTask(IrodsAccessMixin, IrodsBaseTask):
         access_name,
         path,
         user_name,
-        access_lookup,
         irods_backend,
         obj_target=False,
         recursive=True,
@@ -571,7 +577,6 @@ class BatchSetAccessTask(IrodsAccessMixin, IrodsBaseTask):
         access_name,
         paths,
         user_name,
-        access_lookup,
         irods_backend,
         obj_target=False,
         recursive=True,
@@ -584,7 +589,6 @@ class BatchSetAccessTask(IrodsAccessMixin, IrodsBaseTask):
                 access_name,
                 path,
                 user_name,
-                access_lookup,
                 obj_target,
                 recursive,
             )
@@ -595,7 +599,6 @@ class BatchSetAccessTask(IrodsAccessMixin, IrodsBaseTask):
         access_name,
         paths,
         user_name,
-        access_lookup,
         irods_backend,
         obj_target=False,
         recursive=True,
@@ -739,7 +742,6 @@ class BatchMoveDataObjectsTask(IrodsBaseTask):
         src_paths,
         access_name,
         user_name,
-        access_lookup,
         irods_backend,
         *args,
         **kwargs
@@ -788,9 +790,9 @@ class BatchMoveDataObjectsTask(IrodsBaseTask):
             prev_access = None
             if (
                 user_access
-                and user_access.access_name != access_lookup[access_name]
+                and user_access.access_name != ACCESS_LOOKUP[access_name]
             ):
-                prev_access = access_lookup[user_access.access_name]
+                prev_access = ACCESS_LOOKUP[user_access.access_name]
                 modifying_access = True
             elif not user_access:
                 prev_access = 'null'
@@ -823,7 +825,6 @@ class BatchMoveDataObjectsTask(IrodsBaseTask):
         dest_root,
         access_name,
         user_name,
-        access_lookup,
         irods_backend,
         *args,
         **kwargs